requirements.txt
asgiref==3.5.2
Django==4.1.2
django-debug-toolbar==3.7.0
django-filter==22.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
psycopg-binary==3.1.3
psycopg2==2.9.4
pytz==2022.4
sqlparse==0.4.3
